Method: getVariables
I put the getVariables section into it's own function. I kept what you
originally made in the source. Because it create the Map in the main, it
doesn't require a global for now, but that can be changed of course. I
need to know how to pass in the variables from the map into the node for
construction.